Moved some tests to test_suite_formats
