+++ corehq/apps/app_manager/tests/test_suite.py
@@ -1,8 +1,4 @@
-import hashlib
-
 from django.test import SimpleTestCase
-
-import commcare_translations
 
 from corehq.apps.app_manager.exceptions import (
     SuiteValidationError,
@@ -166,204 +162,6 @@ class SuiteTest(SimpleTestCase, TestXmlMixin, SuiteMixin):
 
     def test_case_tile_suite(self, *args):
         self._test_generic_suite("app_case_tiles", "suite-case-tiles")
-
-    def test_case_detail_conditional_enum(self, *args):
-        app = Application.new_app('domain', 'Untitled Application')
-
-        module = app.add_module(Module.new_module('Unititled Module', None))
-        module.case_type = 'patient'
-
-        module.case_details.short.columns = [
-            DetailColumn(
-                header={'en': 'Gender'},
-                model='case',
-                field='gender',
-                format='conditional-enum',
-                enum=[
-                    MappingItem(key="gender = 'male' and age <= 21", value={'en': 'Boy'}),
-                    MappingItem(key="gender = 'female' and age <= 21", value={'en': 'Girl'}),
-                    MappingItem(key="gender = 'male' and age > 21", value={'en': 'Man'}),
-                    MappingItem(key="gender = 'female' and age > 21", value={'en': 'Woman'}),
-                ],
-            ),
-        ]
-
-        key1_varname = hashlib.md5("gender = 'male' and age <= 21".encode('utf-8')).hexdigest()[:8]
-        key2_varname = hashlib.md5("gender = 'female' and age <= 21".encode('utf-8')).hexdigest()[:8]
-        key3_varname = hashlib.md5("gender = 'male' and age > 21".encode('utf-8')).hexdigest()[:8]
-        key4_varname = hashlib.md5("gender = 'female' and age > 21".encode('utf-8')).hexdigest()[:8]
-
-        icon_mapping_spec = """
-        <partial>
-          <template>
-            <text>
-              <xpath function="if(gender = 'male' and age &lt;= 21, $h{key1_varname}, if(gender = 'female' and age &lt;= 21, $h{key2_varname}, if(gender = 'male' and age &gt; 21, $h{key3_varname}, if(gender = 'female' and age &gt; 21, $h{key4_varname}, ''))))">
-                <variable name="h{key4_varname}">
-                  <locale id="m0.case_short.case_gender_1.enum.h{key4_varname}"/>
-                </variable>
-                <variable name="h{key2_varname}">
-                  <locale id="m0.case_short.case_gender_1.enum.h{key2_varname}"/>
-                </variable>
-                <variable name="h{key3_varname}">
-                  <locale id="m0.case_short.case_gender_1.enum.h{key3_varname}"/>
-                </variable>
-                <variable name="h{key1_varname}">
-                  <locale id="m0.case_short.case_gender_1.enum.h{key1_varname}"/>
-                </variable>
-              </xpath>
-            </text>
-          </template>
-        </partial>
-        """.format(
-            key1_varname=key1_varname,
-            key2_varname=key2_varname,
-            key3_varname=key3_varname,
-            key4_varname=key4_varname,
-        )
-        # check correct suite is generated
-        self.assertXmlPartialEqual(
-            icon_mapping_spec,
-            app.create_suite(),
-            './detail[@id="m0_case_short"]/field/template'
-        )
-        # check app strings mapped correctly
-        app_strings = commcare_translations.loads(app.create_app_strings('en'))
-        self.assertEqual(
-            app_strings['m0.case_short.case_gender_1.enum.h{key1_varname}'.format(key1_varname=key1_varname, )],
-            'Boy'
-        )
-        self.assertEqual(
-            app_strings['m0.case_short.case_gender_1.enum.h{key2_varname}'.format(key2_varname=key2_varname, )],
-            'Girl'
-        )
-        self.assertEqual(
-            app_strings['m0.case_short.case_gender_1.enum.h{key3_varname}'.format(key3_varname=key3_varname, )],
-            'Man'
-        )
-        self.assertEqual(
-            app_strings['m0.case_short.case_gender_1.enum.h{key4_varname}'.format(key4_varname=key4_varname, )],
-            'Woman'
-        )
-
-    def test_case_detail_calculated_conditional_enum(self, *args):
-        app = Application.new_app('domain', 'Untitled Application')
-
-        module = app.add_module(Module.new_module('Unititled Module', None))
-        module.case_type = 'patient'
-
-        module.case_details.short.columns = [
-            DetailColumn(
-                header={'en': 'Gender'},
-                model='case',
-                field="if(gender = 'male', 'boy', 'girl')",
-                format='enum',
-                enum=[
-                    MappingItem(key="boy", value={'en': 'Boy'}),
-                    MappingItem(key="girl", value={'en': 'Girl'}),
-                ],
-            ),
-        ]
-
-        icon_mapping_spec = """
-        <partial>
-          <template>
-            <text>
-              <xpath function="if(if(gender = 'male', 'boy', 'girl') = 'boy', $kboy, if(if(gender = 'male', 'boy', 'girl') = 'girl', $kgirl, ''))">
-                <variable name="kboy">
-                  <locale id="m0.case_short.case_if(gender  'male', 'boy', 'girl')_1.enum.kboy"/>
-                </variable>
-                <variable name="kgirl">
-                  <locale id="m0.case_short.case_if(gender  'male', 'boy', 'girl')_1.enum.kgirl"/>
-                </variable>
-              </xpath>
-            </text>
-          </template>
-        </partial>
-        """
-        # check correct suite is generated
-        self.assertXmlPartialEqual(
-            icon_mapping_spec,
-            app.create_suite(),
-            './detail[@id="m0_case_short"]/field/template'
-        )
-        # check app strings mapped correctly
-        app_strings = commcare_translations.loads(app.create_app_strings('en'))
-        self.assertEqual(
-            app_strings["m0.case_short.case_if(gender  'male', 'boy', 'girl')_1.enum.kboy"],
-            'Boy'
-        )
-        self.assertEqual(
-            app_strings["m0.case_short.case_if(gender  'male', 'boy', 'girl')_1.enum.kgirl"],
-            'Girl'
-        )
-
-    def test_case_detail_icon_mapping(self, *args):
-        app = Application.new_app('domain', 'Untitled Application')
-
-        module = app.add_module(Module.new_module('Untitled Module', None))
-        module.case_type = 'patient'
-
-        module.case_details.short.columns = [
-            DetailColumn(
-                header={'en': 'Age range'},
-                model='case',
-                field='age',
-                format='enum-image',
-                enum=[
-                    MappingItem(key='10', value={'en': 'jr://icons/10-year-old.png'}),
-                    MappingItem(key='age > 50', value={'en': 'jr://icons/old-icon.png'}),
-                    MappingItem(key='15%', value={'en': 'jr://icons/percent-icon.png'}),
-                ],
-            ),
-        ]
-
-        key1_varname = '10'
-        key2_varname = hashlib.md5('age > 50'.encode('utf-8')).hexdigest()[:8]
-        key3_varname = hashlib.md5('15%'.encode('utf-8')).hexdigest()[:8]
-
-        icon_mapping_spec = """
-            <partial>
-              <template form="image" width="13%">
-                <text>
-                  <xpath function="if(age = '10', $k{key1_varname}, if(age > 50, $h{key2_varname}, if(age = '15%', $h{key3_varname}, '')))">
-                    <variable name="h{key2_varname}">
-                      <locale id="m0.case_short.case_age_1.enum.h{key2_varname}"/>
-                    </variable>
-                    <variable name="h{key3_varname}">
-                      <locale id="m0.case_short.case_age_1.enum.h{key3_varname}"/>
-                    </variable>
-                    <variable name="k{key1_varname}">
-                      <locale id="m0.case_short.case_age_1.enum.k{key1_varname}"/>
-                    </variable>
-                  </xpath>
-                </text>
-              </template>
-            </partial>
-        """.format(
-            key1_varname=key1_varname,
-            key2_varname=key2_varname,
-            key3_varname=key3_varname,
-        )
-        # check correct suite is generated
-        self.assertXmlPartialEqual(
-            icon_mapping_spec,
-            app.create_suite(),
-            './detail/field/template[@form="image"]'
-        )
-        # check icons map correctly
-        app_strings = commcare_translations.loads(app.create_app_strings('en'))
-        self.assertEqual(
-            app_strings['m0.case_short.case_age_1.enum.h{key3_varname}'.format(key3_varname=key3_varname,)],
-            'jr://icons/percent-icon.png'
-        )
-        self.assertEqual(
-            app_strings['m0.case_short.case_age_1.enum.h{key2_varname}'.format(key2_varname=key2_varname,)],
-            'jr://icons/old-icon.png'
-        )
-        self.assertEqual(
-            app_strings['m0.case_short.case_age_1.enum.k{key1_varname}'.format(key1_varname=key1_varname,)],
-            'jr://icons/10-year-old.png'
-        )
 
     def test_case_tile_pull_down(self, *args):
         app = Application.new_app('domain', 'Untitled Application')
